add thread_pool example and rename old one to test
